# storage/engine: allow SplitKey to be inlined