Move SplitKey to a header so that it can be inlined by the compiler. This
yields a 5% speedup on an MVCCScan benchmark:

	name                                                  old time/op    new time/op    delta
	MVCCScan_RocksDB/rows=10000/versions=1/valueSize=8-8    2.20ms ± 2%    2.10ms ± 2%  -4.43%  (p=0.000 n=10+9)

	name                                                  old speed      new speed      delta
	MVCCScan_RocksDB/rows=10000/versions=1/valueSize=8-8  36.4MB/s ± 2%  38.1MB/s ± 2%  +4.63%  (p=0.000 n=10+9)

	name                                                  old alloc/op   new alloc/op   delta
	MVCCScan_RocksDB/rows=10000/versions=1/valueSize=8-8    1.04MB ± 0%    1.04MB ± 0%    ~     (p=0.986 n=10+10)

	name                                                  old allocs/op  new allocs/op  delta
	MVCCScan_RocksDB/rows=10000/versions=1/valueSize=8-8      3.00 ± 0%      3.00 ± 0%    ~     (all equal)

Release note: None

Based on a late-night hacking session with @jordanlewis and @nvanbenschoten.